Feedback submission features changed

+++ app/src/main/java/universal/appfactory/aeroindia2023/ZonalManagerAdapter.kt
@@ -34,8 +34,8 @@ class ZonalManagerAdapter(private val ZList: ArrayList<ZonalManagerModel>,privat
         var time:String=DT.substring(11)
         holder.date.text=date
         holder.time.text=time
-        if(itemsViewModel.getstatus()==1)
-        {holder.status.text="Pending"
+        if(itemsViewModel.getstatus().toInt() == 1){
+            holder.status.text="Pending"
             holder.status.setOnClickListener{
                 val intent=Intent(mContext,ResolvedRemarksActivity::class.java)
                 intent.putExtra("CompId",itemsViewModel.getComplaintId())
